Extract inline provisioner scripts to external files
This commit extracts the inline calls to `gcloud` to external scripts,
and replaces the `gcloud `service account activation with a credentials
override that doesn't need setup. Extracting these scripts allows us to
run linters on the scripts, call and test them independently, and
reduces the cost of development on those scripts.