Configure credentials override for `gcloud` in tests
In order to use `gcloud` for our test verification we need to provide
credentials. This commit amends the inspec tests to set the
`CLOUDSDK_AUTH_CREDENTIAL_FILE_OVERRIDE` environment variable inside of
the tests to point at the module credentials, as given by the
`TF_VAR_credentials_file_path` variable.

Note that this approach is pretty heavy-handed; if someone wanted to
manage the value of `CLOUDSDK_AUTH_CREDENTIAL_FILE_OVERRIDE` themselves
this change will ignore the user set value. If this becomes a problem we
can investigate more nuanced solutions.